test(wallet): add helpers to `test_utils`
